requirements.txt
streamlit
pandas
nltk
scikit-learn
docx2txt
PyPDF2
plotly
textrazor
python-dotenv
